refactor(engine): move StreamProcessor and state machines to streamprocessors package
