+++ engine/src/main/java/io/camunda/zeebe/engine/api/ReplayContext.java
@@ -7,6 +7,4 @@
  */
 package io.camunda.zeebe.engine.api;
 
-public interface ReplayContext {
-
-}
+public interface ReplayContext {}
