+++ engine/src/main/java/io/camunda/zeebe/engine/processing/streamprocessor/MetadataFilter.java
@@ -8,6 +8,7 @@
 package io.camunda.zeebe.engine.processing.streamprocessor;
 
 import io.camunda.zeebe.protocol.impl.record.RecordMetadata;
+import io.camunda.zeebe.streamprocessor.StreamProcessor;
 import java.util.Objects;
 
 /**
