+++ engine/src/main/java/io/camunda/zeebe/engine/processing/streamprocessor/StreamProcessorListener.java
@@ -8,6 +8,7 @@
 package io.camunda.zeebe.engine.processing.streamprocessor;
 
 import io.camunda.zeebe.logstreams.log.LoggedEvent;
+import io.camunda.zeebe.streamprocessor.StreamProcessor;
 
 /**
  * A listener for the {@link StreamProcessor}. Allows retrieving insides of the processing and
